refactoring cameras and scene
